rename FlyLight to LightSpace and fix test dependency
