+++ package.json
@@ -43,9 +43,9 @@
     "postcss-flexbugs-fixes": "3.0.0",
     "postcss-loader": "2.0.5",
     "promise": "7.1.1",
-    "react-addons-test-utils": "^15.6.0",
     "react-dev-utils": "^3.0.0",
     "react-error-overlay": "^1.0.7",
+    "react-test-renderer": "^15.6.1",
     "redux-devtools-extension": "^2.13.2",
     "style-loader": "0.17.0",
     "sw-precache-webpack-plugin": "0.9.1",
