+++ src/components/light/Light.test.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import Light from './Light';
+
+import { shallow } from 'enzyme';
+
+describe('Light', () => {
+	it('', () => {
+
+	});
+});
